modify tests, moves interpreter
